fix crash on inputting integer value
